changed deletion of asset groups to allow service users to delete

+++ src/main/java/dk/northtech/dasscoassetservice/services/AssetGroupService.java
@@ -115,6 +115,8 @@ public class AssetGroupService {
             throw new IllegalArgumentException("Asset group does not exist!");
         }
 
+        rightsValidationService.checkAssetGroupOwnershipThrowing(user, assetGroupOptional.get());
+
         jdbi.onDemand(AssetGroupRepository.class).deleteAssetGroup(groupName.toLowerCase(), user);
     }
